requirements.txt
fastapi
uvicorn
python-multipart
python-docx
pandas
PyMuPDF
